Remove `udpPort` CLI option from waku v1 and v2 nodes
The UDP port is merely used to satisfy underlying `nim-eth` APIs
(namely `proc redirectPorts` which expects a UDP port).

Since there's no support for UDP transport in nim-waku, there's no
need to have a CLI option for that.

This commit removes the option from the CLI while keeping the value
inside application code in tact to not break calls to third-party
APIs.